+++ NEWS.md
@@ -3,6 +3,7 @@
 * DESCRIPTION: Updated.
 * Added Imports for `manipulate` and `relax`.  Added Imports for `graphics`, `grDevices`, `stats`, and `utils`.  Added Suggests for `tcltk`.
 * `growthModelSim()`: Modified.  Fixed links to functions in `FSA` that had changed names.
+* `mrClosed1Sim()`: Modified.  Changed to use `manipulate` rather than `relax`.  Added the `sim=` argument which allows simulations of the `distribution` or for assessing `assumptions` violations.  Streamlined code some.  Changed look of the histogram (more like `hist()` from `FSA`).
 * `vbComp()`: Modified.  Reorganized code and deleted bad fishR links in documentation.
 * `vbDataGen()`: Modified.  Removed bad link to `gReshape()` in `FSA`, which no longer exists.